More updates to supervised tasks workflow

+++ analyses/scripts/R/fit_mofa.R
@@ -37,7 +37,7 @@ option_list <- list(
     make_option("--output_dim", type="integer", default=10, help="Number of dimensions the output should take"),
     make_option("--omic_types", type="character", default="mutation:methylation:mrnaseq:cna", help="List of omic assays to use, separated by colons (no spaces). Default 'mrnaseq:methylation:mutation:cna'."),
     make_option("--is_grouped", action="store_true", default=TRUE, help="Toggles sample grouping by cancer type."),
-    make_option("--var_filter", type="numeric", default=0.5, help="fraction of most-variable features to keep within each view"),
+    make_option("--var_filter", type="double", default=0.5, help="fraction of most-variable features to keep within each view"),
     make_option("--mofa_python", help="path to python that has mofapy2 installed")
     )
